% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.getShuffleWorkspace}
\alias{.getShuffleWorkspace}
\title{Get a cached ShuffleWorkspace XPtr for the given (d, depth).}
\usage{
.getShuffleWorkspace(d, depth)
}
\description{
Get a cached ShuffleWorkspace XPtr for the given (d, depth).
}
\keyword{internal}
