% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.getWorkspace}
\alias{.getWorkspace}
\title{Get a cached SigWorkspace XPtr for the given (d, depth).}
\usage{
.getWorkspace(d, depth)
}
\description{
Get a cached SigWorkspace XPtr for the given (d, depth).
}
\keyword{internal}
