% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.inferDim}
\alias{.inferDim}
\title{Infer alphabet size d from a (d, depth) signature length.}
\usage{
.inferDim(L, depth)
}
\description{
Infer alphabet size d from a (d, depth) signature length.
}
\keyword{internal}
