% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lyndonWords.R
\name{.lexLess}
\alias{.lexLess}
\title{Strict lex-less comparison of equal-length integer vectors.}
\usage{
.lexLess(a, b)
}
\description{
Strict lex-less comparison of equal-length integer vectors.
}
\keyword{internal}
