% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.lyndonWordNames}
\alias{.lyndonWordNames}
\title{Build Lyndon-word names with the given separator.}
\usage{
.lyndonWordNames(words, sep = ",")
}
\description{
Build Lyndon-word names with the given separator.
}
\keyword{internal}
