% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{.tensorExpRef}
\alias{.tensorExpRef}
\title{Pure-R reference truncated tensor exponential.}
\usage{
.tensorExpRef(a, depth)
}
\description{
\code{exp(a)} = sum over \code{k} from 0 to \code{N} of \verb{a^(x k) / k!}. Used in tests.
}
\keyword{internal}
