% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.validateSignatureOptions}
\alias{.validateSignatureOptions}
\title{Validate the shared \code{sep} / \code{includeLevelZero} options.}
\usage{
.validateSignatureOptions(sep, includeLevelZero)
}
\description{
Validate the shared \code{sep} / \code{includeLevelZero} options.
}
\keyword{internal}
